+++ main/gbd.py
@@ -6,11 +6,11 @@ import os
 import argparse
 import re
 
-from core.main import tags, groups, search, import_data
+from main.core.main import groups, import_data, search, tags
 
-from core.database.db import Database
-from core.hashing.gbd_hash import gbd_hash
-from core.main.util import eprint, read_hashes, confirm
+from main.core.database.db import Database
+from main.core.hashing.gbd_hash import gbd_hash
+from main.core.main.util import eprint, read_hashes, confirm
 
 from os.path import realpath, dirname, join
